# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "帖子服务，提供帖子发布、查询、管理等功能。",
        "title": "Post Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "paths": {
        "/api/v1/post/admin/posts": {
            "get": {
                "description": "出于管理目的，根据各种过滤条件检索分页的帖子列表。使用查询参数进行过滤和分页。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-posts (管理员-帖子)"
                ],
                "summary": "按条件列出帖子 (管理员)",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "uint64",
                        "description": "按精确的帖子 ID 过滤",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "按帖子标题过滤（模糊匹配）",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "按作者用户名过滤（模糊匹配）",
                        "name": "author_username",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1,
                            2
                        ],
                        "type": "integer",
                        "description": "按帖子状态过滤 (0=待审核, 1=已审核, 2=已拒绝)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1,
                            2,
                            3
                        ],
                        "type": "integer",
                        "description": "按官方标签过滤 (例如, 0=无, 1=官方认证)",
                        "name": "official_tag",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "按最小浏览量过滤",
                        "name": "view_count_min",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "按最大浏览量过滤",
                        "name": "view_count_max",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "created_at",
                            "updated_at"
                        ],
                        "type": "string",
                        "default": "created_at",
                        "description": "排序字段 (created_at 或 updated_at)",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "是否降序排序 (true 为 DESC, false/省略为 ASC)",
                        "name": "order_desc",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "format": "int",
                        "description": "页码（从 1 开始）",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "format": "int",
                        "description": "每页帖子数量",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "帖子检索成功\" // \u003c--- 修改",
                        "schema": {
                            "$ref": "#/definitions/vo.ListPostsAdminResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "无效的输入参数（例如，无效的 page, page_size, status）\" // \u003c--- 修改",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "检索帖子时发生内部服务器错误\" // \u003c--- 修改",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    }
                }
            }
        },
...
        "/api/v1/post/admin/posts/{post_id}": {
            "delete": {
                "description": "管理员软删除指定ID的帖子 (Admin soft deletes a post with the specified ID)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "管理员删除帖子 (Admin delete post)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "帖子ID (Post ID)",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "帖子删除成功",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "无效的帖子ID格式",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "401": {
                        "description": "管理员未登录或无权限",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "404": {
                        "description": "帖子未找到",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "删除帖子时发生内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    }
                }
            }
        },
...
        "/api/v1/post/posts": {
            "post": {
                "description": "使用提供的详情（作为独立表单字段）和图片文件创建一个新帖子。请求体应为 multipart/form-data。",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts (帖子)"
                ],
                "summary": "创建新帖子 (独立表单字段及图片)",
                "parameters": [
                    {
                        "maxLength": 100,
                        "type": "string",
                        "description": "帖子标题",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "maxLength": 1000,
                        "type": "string",
                        "description": "帖子内容",
                        "name": "content",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "minimum": 0,
                        "type": "number",
                        "description": "单价 (可选, 大于等于0)",
                        "name": "price_per_unit",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "联系方式 (可选)",
                        "name": "contact_info",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "作者ID",
                        "name": "author_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "url",
                        "description": "作者头像 URL (可选, 需为有效URL)",
                        "name": "author_avatar",
                        "in": "formData"
                    },
                    {
                        "maxLength": 50,
                        "type": "string",
                        "description": "作者用户名",
                        "name": "author_username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "帖子图片文件 (可多选)",
                        "name": "images",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "帖子创建成功",
                        "schema": {
                            "$ref": "#/definitions/vo.PostDetailResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "无效的请求负载或文件处理错误",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "创建帖子时发生内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    }
                }
            }
        },
        "/api/v1/post/posts/by-author": {
            "get": {
                "description": "使用游标分页方式，检索特定用户公开发布的帖子列表。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts (帖子)"
                ],
                "summary": "获取指定用户的帖子列表 (公开, 游标加载)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "要查询其帖子的用户 ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "uint64",
                        "description": "游标（上一页最后一个帖子的 ID），首页省略",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "format": "int",
                        "description": "每页帖子数量",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "帖子检索成功\" // 确保 vo.ListPostsByUserIDResponseWrapper 对应游标加载的响应结构",
                        "schema": {
                            "$ref": "#/definitions/vo.ListPostsByCursorResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "无效的输入参数",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "检索帖子时发生内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    }
                }
            }
        },
        "/api/v1/post/posts/mine": {
            "get": {
                "description": "获取当前登录用户发布的帖子列表，支持按官方标签、标题、帖子状态筛选，并使用分页加载。UserID 从请求上下文中获取。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts (帖子)"
                ],
                "summary": "获取我的帖子列表",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "format": "int32",
                        "default": 1,
                        "description": "页码 (从1开始)",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "format": "int32",
                        "default": 10,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            0,
                            1,
                            2,
                            3
                        ],
                        "type": "integer",
                        "format": "int32",
                        "description": "官方标签 (0:无标签, 1:官方认证, 2:预付保证金, 3:急速响应)",
                        "name": "officialTag",
                        "in": "query"
                    },
                    {
                        "maxLength": 255,
                        "type": "string",
                        "description": "标题模糊搜索关键词 (最大长度 255)",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1,
                            2
                        ],
                        "type": "integer",
                        "format": "int32",
                        "description": "帖子状态 (0:待审核, 1:审核通过, 2:拒绝)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应，包含用户帖子列表和总记录数",
                        "schema": {
                            "$ref": "#/definitions/vo.ListUserPostPageResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "无效的请求参数",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "401": {
                        "description": "用户未授权或认证失败",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    }
                }
            }
        },
        "/api/v1/post/posts/timeline": {
            "get": {
                "description": "根据指定条件（官方标签、标题、作者用户名）和游标分页获取帖子列表，按时间倒序排列。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts (帖子)"
                ],
                "summary": "获取帖子时间线列表 (公开)",
                "parameters": [
                    {
                        "type": "string",
                        "format": "date-time",
                        "description": "上一页最后一条记录的创建时间 (RFC3339格式, e.g., 2023-01-01T15:04:05Z)",
                        "name": "lastCreatedAt",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "format": "uint64",
                        "description": "上一页最后一条记录的帖子ID",
                        "name": "lastPostId",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "format": "int32",
                        "default": 10,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            0,
                            1,
                            2,
                            3
                        ],
                        "type": "integer",
                        "format": "int32",
                        "description": "官方标签 (0:无标签, 1:官方认证, 2:预付保证金, 3:急速响应)",
                        "name": "officialTag",
                        "in": "query"
                    },
                    {
                        "maxLength": 255,
                        "type": "string",
                        "description": "标题模糊搜索关键词 (最大长度 255)",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "maxLength": 50,
                        "type": "string",
                        "description": "作者用户名模糊搜索关键词 (最大长度 50)",
                        "name": "authorUsername",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应，包含帖子列表和下一页游标信息",
                        "schema": {
                            "$ref": "#/definitions/vo.PostTimelinePageResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "无效的请求参数",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "vo.BaseResponseWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "成功时为 0, 错误时为具体错误码",
                    "type": "integer",
                    "example": 0
                },
                "message": {
                    "description": "成功或错误消息",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "vo.ListPostsAdminByConditionResponse": {
            "type": "object",
            "properties": {
                "posts": {
                    "description": "帖子列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.PostResponse"
                    }
                },
                "total": {
                    "description": "帖子总数",
                    "type": "integer"
                }
            }
        },
        "vo.ListPostsAdminResponseWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "使用具体的 vo.ListPostsAdminByConditionResponse",
                    "allOf": [
                        {
                            "$ref": "#/definitions/vo.ListPostsAdminByConditionResponse"
                        }
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "vo.ListPostsByCursorResponseWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "使用具体的 vo.ListHotPostsByCursorResponse",
                    "allOf": [
                        {
                            "$ref": "#/definitions/vo.ListHotPostsByCursorResponse"
                        }
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "vo.PostDetailResponseWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/vo.PostDetailVO"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
    }
}